+++ vimiv/commands/search.py
@@ -57,6 +57,7 @@ class Search(QObject):
         super().__init__()
         self._text = ""
         self._reverse = False
+        api.signals.cancel.connect(self.clear)
 
     def __call__(
         self, text, mode, count=0, reverse=False, incremental=False