+++ docs/changelog.rst
@@ -16,6 +16,9 @@ Added:
   * ``6`` or ``PyQt6``: Use PyQt6.
   * ``PySide6``: Use PySide6 (Qt for Python). This is highly experimental and should be
     used with care.
+* The generic ``:cancel`` command bound to ``<escape>`` together with the
+  ``api.signals.cancel`` signal which can be used by plugins in case they need some sort
+  of resetting.
 
 Fixed:
 ^^^^^^